Add files via upload
It seems that windows or git is doing something weird with the images. In previous commits that files would end with .JPG instead of .jpg which would cause them not to load on the production site. I have just used the option to add files on GitHub. The preview looks like it is the correct file type so I think this should resolve the issue.